backend/requirements.txt
beautifulsoup4
requests
flask
flask-sqlalchemy
flask-cors
python-dotenv
psycopg2-binary
mcp
